* scroll.c (do_scrolling): When bcopying the max_ascent field from 	current_frame to temp_frame, remember that max_ascent is an array 	of shorts, not ints.

+++ src/scroll.c
@@ -265,7 +265,7 @@ do_scrolling (frame, matrix, window_size, unchanged_at_top)
       bcopy (current_frame->pix_height, temp_frame->pix_height,
 	     current_frame->height * sizeof (short));
       bcopy (current_frame->max_ascent, temp_frame->max_ascent,
-	     current_frame->height * sizeof (int));
+	     current_frame->height * sizeof (short));
     }
 #endif
